chore: rename "dispatcher" to "ansible-dispatcher"
