+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -15,9 +15,9 @@ body:
       options:
         - label: I agree to follow this project's [code of conduct](https://docs.ansible.com/ansible/latest/community/code_of_conduct.html).
           required: true
-        - label: I have checked the [current issues](https://github.com/ansible/dispatcher/issues) for duplicates.
+        - label: I have checked the [current issues](https://github.com/ansible/ansible-dispatcher/issues) for duplicates.
           required: true
-        - label: I understand that dispatcher is open source software provided for free and that I might not receive a timely response.
+        - label: I understand that ansible-dispatcher is open source software provided for free and that I might not receive a timely response.
           required: true
 
   - type: textarea
