+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,12 +1,12 @@
 blank_issues_enabled: false
 contact_links:
   - name: For debugging help or technical support
-    url: https://github.com/ansible/dispatcher/blob/main/docs/contributing.md#contact
+    url: https://github.com/ansible/ansible-dispatcher/blob/main/docs/contributing.md#contact
     about: For general debugging or technical support please see the linked section of our docs.
 
   - name: 📝 Ansible Code of Conduct
     url: https://docs.ansible.com/ansible/latest/community/code_of_conduct.html
-    about: Dispatcher uses the Ansible Code of Conduct; ❤ Be nice to other members of the community. ☮ Behave.
+    about: Ansible Dispatcher uses the Ansible Code of Conduct; ❤ Be nice to other members of the community. ☮ Behave.
 
   - name: 💼 For Enterprise
     url: https://www.redhat.com/en/technologies/management/ansible
