+++ .github/workflows/ci.yml
@@ -23,7 +23,7 @@ jobs:
         with:
           python-version: ${{ matrix.py_version }}
 
-      - name: Install dispatcher
+      - name: Install ansible_dispatcher
         run: pip install -e .[pg_notify]
       - run: make postgres
       - run: pip install pytest pytest-asyncio
@@ -38,7 +38,7 @@ jobs:
           show-progress: false
 
       - run: pip install black
-      - run: black --check dispatcher
+      - run: black --check ansible_dispatcher
 
   isort:
     name: Run isort
@@ -49,7 +49,7 @@ jobs:
           show-progress: false
 
       - run: pip install isort
-      - run: isort --check dispatcher
+      - run: isort --check ansible_dispatcher
 
   mypy:
     name: Run mypy
@@ -62,7 +62,7 @@ jobs:
       - run: pip install mypy
       - run: pip install -e .[pg_notify]
       - run: python3 -m pip install types-PyYAML
-      - run: mypy --ignore-missing-imports dispatcher
+      - run: mypy --ignore-missing-imports ansible_dispatcher
 
   flake8:
     name: Run flake8
@@ -73,4 +73,4 @@ jobs:
           show-progress: false
 
       - run: pip install flake8
-      - run: flake8 dispatcher
+      - run: flake8 ansible_dispatcher
