+++ ansible_dispatcher/main.py
@@ -5,11 +5,11 @@ import signal
 from types import SimpleNamespace
 from typing import Optional, Union
 
-from dispatcher.pool import WorkerPool
-from dispatcher.producers.base import BaseProducer
-from dispatcher.producers.brokered import BrokeredProducer
-from dispatcher.producers.scheduled import ScheduledProducer
-from dispatcher.utils import MODULE_METHOD_DELIMITER
+from ansible_dispatcher.pool import WorkerPool
+from ansible_dispatcher.producers.base import BaseProducer
+from ansible_dispatcher.producers.brokered import BrokeredProducer
+from ansible_dispatcher.producers.scheduled import ScheduledProducer
+from ansible_dispatcher.utils import MODULE_METHOD_DELIMITER
 
 logger = logging.getLogger(__name__)
 
@@ -217,7 +217,7 @@ class DispatcherMain:
                 self.control_count += 1
                 await self.pool.dispatch_task(
                     {
-                        'task': f'dispatcher.brokers.{broker.broker}{MODULE_METHOD_DELIMITER}publish_message',
+                        'task': f'ansible_dispatcher.brokers.{broker.broker}{MODULE_METHOD_DELIMITER}publish_message',
                         'args': [message['reply_to'], json.dumps(returned)],
                         'kwargs': {'config': broker.config, 'new_connection': True},
                         'uuid': f'control-{self.control_count}',
