+++ ansible_dispatcher/producers/base.py
