+++ docs/roadmap.md
@@ -28,8 +28,9 @@ We want it to be _impossible_ to run a task with parameters
 other than what are declared on `@task`
 
 Think of it this way, there are
- - runtime arguments (or parameters)
- - configuration parameters
+
+- runtime arguments (or parameters)
+- configuration parameters
 
 Something like a task timeout is a configuration parameter.
 This implies that when we register a method with `@task`
@@ -48,8 +49,9 @@ detailed tracking timing for every time a task is called.
 This goes with the call, not the task.
 
 So the new types we probably want are:
- - `Task`
- - `Call`
+
+- `Task`
+- `Call`
 
 This is in addition to the `PoolWorker` which is the worker that
 runs the task.
@@ -74,28 +76,29 @@ mostly scalar type data of floats, strings, ids, etc.
 
 #### Finish integrating publisher logic
 
-The content existing in `dispatcher.publish` is mostly not connected.
+The content existing in `ansible_dispatcher.publish` is mostly not connected.
 
-What's interesting here is that `dispatcher.publish` should import
+What's interesting here is that `ansible_dispatcher.publish` should import
 from the broker module.
 That gets hard to manage with multiple connections (ala Django).
 But some version of it we should do...
 
 #### Finish integrating the worker loop
 
-Overlapping with the publisher stuff, the `dispatcher.publish` should
+Overlapping with the publisher stuff, the `ansible_dispatcher.publish` should
 get the method name, and the args, import the method, and run it.
 
 This requires moving more code in from AWX
 
-https://github.com/ansible/awx/blob/devel/awx/main/dispatch/worker/task.py
+<https://github.com/ansible/awx/blob/devel/awx/main/dispatch/worker/task.py>
 
 That has
- - importing logic
- - calling logic
- - supporting stuff to include timings
- - signal handling
- - exception handling
+
+- importing logic
+- calling logic
+- supporting stuff to include timings
+- signal handling
+- exception handling
 
 ### Post-alpha
 
@@ -120,7 +123,7 @@ Make AWX run using this library, this should be an early goal in this stage.
 
 A moderate version of this was proposed in:
 
-https://github.com/ansible/awx/pull/14749
+<https://github.com/ansible/awx/pull/14749>
 
 In grand conclusion, there is no way to assure that the LISTSEN connection
 is not dropped.
@@ -134,7 +137,7 @@ by experiential means.
 To know if a connection works, you must publish a control message and receive it.
 To know if a worker is alive, you must send a message and receive a reply.
 
-Because of this knowledge, the new dispatcher library must just straight to this eventuality.
+Because of this knowledge, the new ansible_dispatcher library must just straight to this eventuality.
 Implement checks for brokers and workers based on send-and-receive.
 This can be done fully with asynio patterns.
 
@@ -147,16 +150,18 @@ Several very practical problems are not intended to ever be solved by the dispat
 However, for someone using postgres or any other modern database,
 combined with the dispatcher, they have the ability to solve these problems.
 
-https://github.com/ansible/awx/issues/11997
+<https://github.com/ansible/awx/issues/11997>
 
 Breakdown of those problems:
+
 1. Have a node in the cluster, any node, process a task
 2. Have a periodic task run, anywhere in the cluster, at a certain frequency
 
 The solution for (1) is to add an entry to a table when submitting the task.
 Then depending on the use case, there are 2 decent options:
- - broadcast a task asking any willing node to run the task, get lock, if lock is taken, bail
- - run a periodic task that will use `select_for_update` to get entries and mark as received
+
+- broadcast a task asking any willing node to run the task, get lock, if lock is taken, bail
+- run a periodic task that will use `select_for_update` to get entries and mark as received
 
 The solution for (2) in AWX uses the Solo model to track a `datetime`.
 This is self-obviously needed for the feature of _user_ schedules.
@@ -167,7 +172,7 @@ When using `@task()` decorator, we add `timeout=5` to timeout in 5 seconds.
 
 A solution was drafted in the branch:
 
-https://github.com/ansible/awx/compare/devel...AlanCoding:awx:dispatcher_timeout
+<https://github.com/ansible/awx/compare/devel...AlanCoding:awx:dispatcher_timeout>
 
 #### Singleton Tasks
 
@@ -176,8 +181,9 @@ but a more efficient alternative is to never start those tasks.
 
 This proposes another argument to `@task()` decorator that makes the task exclusive.
 When another version of the task is already running, there are 2 sub-options we could do:
- - wait for the existing task to finish before running the new task
- - discard the new task
+
+- wait for the existing task to finish before running the new task
+- discard the new task
 
 The use cases for AWX mainly wand the 2nd one.
 Idepotent tasks are used extremely heavily on schedules, meaning that
