+++ tests/unit/test_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from dispatcher.utils import resolve_callable
+from ansible_dispatcher.utils import resolve_callable
 
 
 def test_resolve_lamda_method():
